requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pyjwt==2.10.1
pyyaml==6.0.2
python-multipart==0.0.20
aiofiles==24.1.0
bcrypt==4.2.0
python-dotenv==1.0.1
